dw: add picture and update
build farm picture
remove long log
remove debug for watch
change email to
add temp email

+++ ihook/watch.sh
@@ -126,5 +126,5 @@ do
 done
 
 rm -f $WATCHDOG_PATH/tmp
-rm -fr $TASK_SPACE/$WATCH_TMP
-
+[[ -z $DEBUG ]] && rm -fr $TASK_SPACE/$WATCH_TMP
+[[ -z $DEBUG ]] && rm -fr $TASK_SPACE/$ISPEC_TMP
